+++ lib/src/feature/widget/widget/dropdown/selection_dropdown/view_model/myoro_single_selection_dropdown_view_model.dart
@@ -9,6 +9,7 @@ final class MyoroSingleSelectionDropdownViewModel<T>
     : super(
         MyoroSingleInputDropdownController<T>(
           configuration: MyoroSingleInputDropdownConfiguration(
+            label: configuration.label,
             dropdownType: configuration.dropdownType,
             menuConfiguration: configuration.menuConfiguration,
             selectedItemBuilder: configuration.selectedItemBuilder,